host: fix broken pip installation due to the README file
Rename README to README.md, which is both consistent with how it is in GreatFET, as well as what the setup.py expects.
Also change setup.py to expect README.md to be the the `host` directory, instead of its parent directory.

`pip install host/` was previously failing as it couldn't find the expected README.md file.